+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 The changelog was created to keep track of most important changes between different stages of the competition (Hindcast Stage, Forecast Stage, Final Prize Stage).
 
 # Changes between Forecast Stage and Final Prize Stage
+* added new features:
+	* PREC_DAILY_prev (previous available value of PREC_DAILY from SNOTEL),
+	* PREC_DAILY_Apr_mean (average value since April of PREC_DAILY from SNOTEL),
+	* PREC_DAILY_Apr_prev_diff (difference between previous available value of PREC_DAILY and a value from Mar 30 from SNOTEL),
+	* WTEQ_DAILY_Jun_prev_diff (difference between previous available value of WTEQ_DAILY and a value from May 30 from SNOTEL),
+	* WTEQ_DAILY_Jul_prev_diff (difference between previous available value of WTEQ_DAILY and a value from Jun 29 from SNOTEL),
+	* sd_prev (previous available value of Snow depth from CDS)
 * added volume residuals prediction for May-Jul models for 23 site ids (new function lgbm_cv_residuals)
 * added May-Jul predictions for 3 site ids without nat flow previous values from a given water year, model is still built for all site ids but only 3 are evaluated (incorporated into lgbm_cv function)
 * added better DISTR_PERC for February and March predictions